+++ libs/extractor/src/extractor/extract_global_style_from_expression.rs
@@ -31,53 +31,56 @@ pub fn extract_global_style_from_expression<'a>(
 
     if let Expression::ObjectExpression(obj) = expression {
         for p in obj.properties.iter_mut() {
-            if let ObjectPropertyKind::ObjectProperty(o) = p
-                && let Some(name) = if let PropertyKey::StaticIdentifier(ident) = &o.key {
-                    Some(ident.name.to_string())
-                } else if let PropertyKey::StringLiteral(s) = &o.key {
-                    Some(s.value.to_string())
-                } else if let PropertyKey::TemplateLiteral(t) = &o.key {
-                    Some(
-                        t.quasis
-                            .iter()
-                            .map(|q| q.value.raw.as_str())
-                            .collect::<String>(),
-                    )
-                } else {
-                    None
-                }
-            {
-                if name == "imports" {
-                    if let Expression::ArrayExpression(arr) = &o.value {
-                        for p in arr.elements.iter() {
-                            if let Some(url) = if let ArrayExpressionElement::StringLiteral(s) = p {
-                                Some(s.value.trim().to_string())
-                            } else if let ArrayExpressionElement::TemplateLiteral(t) = p {
-                                Some(
-                                    t.quasis
-                                        .iter()
-                                        .map(|q| q.value.raw.as_str())
-                                        .collect::<String>()
-                                        .trim()
-                                        .to_string(),
-                                )
-                            } else {
-                                None
-                            } {
-                                styles.push(ExtractStyleProp::Static(ExtractStyleValue::Import(
-                                    ExtractImport {
-                                        url,
-                                        file: file.to_string(),
-                                    },
-                                )));
+            match p {
+                ObjectPropertyKind::ObjectProperty(o) => {
+                    if let Some(name) = if let PropertyKey::StaticIdentifier(ident) = &o.key {
+                        Some(ident.name.to_string())
+                    } else if let PropertyKey::StringLiteral(s) = &o.key {
+                        Some(s.value.to_string())
+                    } else if let PropertyKey::TemplateLiteral(t) = &o.key {
+                        Some(
+                            t.quasis
+                                .iter()
+                                .map(|q| q.value.raw.as_str())
+                                .collect::<String>(),
+                        )
+                    } else {
+                        None
+                    } {
+                        if name == "imports" {
+                            if let Expression::ArrayExpression(arr) = &o.value {
+                                for p in arr.elements.iter() {
+                                    if let Some(url) =
+                                        if let ArrayExpressionElement::StringLiteral(s) = p {
+                                            Some(s.value.trim().to_string())
+                                        } else if let ArrayExpressionElement::TemplateLiteral(t) = p
+                                        {
+                                            Some(
+                                                t.quasis
+                                                    .iter()
+                                                    .map(|q| q.value.raw.as_str())
+                                                    .collect::<String>()
+                                                    .trim()
+                                                    .to_string(),
+                                            )
+                                        } else {
+                                            None
+                                        }
+                                    {
+                                        styles.push(ExtractStyleProp::Static(
+                                            ExtractStyleValue::Import(ExtractImport {
+                                                url,
+                                                file: file.to_string(),
+                                            }),
+                                        ));
+                                    }
+                                }
                             }
-                        }
-                    }
-                } else if name == "fontFaces" {
-                    if let Expression::ArrayExpression(arr) = &o.value {
-                        for p in arr.elements.iter() {
-                            if let ArrayExpressionElement::ObjectExpression(o) = p {
-                                styles.push(ExtractStyleProp::Static(ExtractStyleValue::FontFace(ExtractFontFace {
+                        } else if name == "fontFaces" {
+                            if let Expression::ArrayExpression(arr) = &o.value {
+                                for p in arr.elements.iter() {
+                                    if let ArrayExpressionElement::ObjectExpression(o) = p {
+                                        styles.push(ExtractStyleProp::Static(ExtractStyleValue::FontFace(ExtractFontFace {
                                     properties: BTreeMap::from_iter(
                                         o.properties
                                             .iter()
@@ -103,54 +106,66 @@ pub fn extract_global_style_from_expression<'a>(
                                     ),
                                     file: file.to_string(),
                                 })));
-                            } else if let ArrayExpressionElement::TemplateLiteral(t) = p {
-                                let css_styles = css_to_style(
-                                    t.quasis
-                                        .iter()
-                                        .map(|q| q.value.raw.as_str())
-                                        .collect::<String>()
-                                        .trim(),
-                                    0,
-                                    &None,
-                                )
-                                .into_iter()
-                                .map(ExtractStyleValue::Static)
-                                .collect::<Vec<_>>();
-                                styles.push(ExtractStyleProp::Static(ExtractStyleValue::FontFace(
-                                    ExtractFontFace {
-                                        properties: BTreeMap::from_iter(
-                                            css_styles.iter().filter_map(|p| {
-                                                if let ExtractStyleValue::Static(st) = p {
-                                                    Some((
-                                                        st.property().to_string(),
-                                                        st.value().to_string(),
-                                                    ))
-                                                } else {
-                                                    None
-                                                }
+                                    } else if let ArrayExpressionElement::TemplateLiteral(t) = p {
+                                        let css_styles = css_to_style(
+                                            t.quasis
+                                                .iter()
+                                                .map(|q| q.value.raw.as_str())
+                                                .collect::<String>()
+                                                .trim(),
+                                            0,
+                                            &None,
+                                        )
+                                        .into_iter()
+                                        .map(ExtractStyleValue::Static)
+                                        .collect::<Vec<_>>();
+                                        styles.push(ExtractStyleProp::Static(
+                                            ExtractStyleValue::FontFace(ExtractFontFace {
+                                                properties: BTreeMap::from_iter(
+                                                    css_styles.iter().filter_map(|p| {
+                                                        if let ExtractStyleValue::Static(st) = p {
+                                                            Some((
+                                                                st.property().to_string(),
+                                                                st.value().to_string(),
+                                                            ))
+                                                        } else {
+                                                            None
+                                                        }
+                                                    }),
+                                                ),
+                                                file: file.to_string(),
                                             }),
-                                        ),
-                                        file: file.to_string(),
-                                    },
-                                )));
+                                        ));
+                                    }
+                                }
                             }
+                        } else {
+                            styles.extend(
+                                extract_style_from_expression(
+                                    ast_builder,
+                                    None,
+                                    &mut o.value,
+                                    0,
+                                    &Some(StyleSelector::Global(
+                                        if let Some(name) = name.strip_prefix("_") {
+                                            StyleSelector::from(name).to_string().replace("&", "*")
+                                        } else {
+                                            name.to_string()
+                                        },
+                                        file.to_string(),
+                                    )),
+                                )
+                                .styles,
+                            );
                         }
                     }
-                } else {
+                }
+                ObjectPropertyKind::SpreadProperty(o) => {
                     styles.extend(
-                        extract_style_from_expression(
+                        extract_global_style_from_expression(
                             ast_builder,
-                            None,
-                            &mut o.value,
-                            0,
-                            &Some(StyleSelector::Global(
-                                if let Some(name) = name.strip_prefix("_") {
-                                    StyleSelector::from(name).to_string().replace("&", "*")
-                                } else {
-                                    name.to_string()
-                                },
-                                file.to_string(),
-                            )),
+                            o.argument.get_inner_expression_mut(),
+                            file,
                         )
                         .styles,
                     );