+++ packages/components/src/components/Input/index.tsx
@@ -17,10 +17,12 @@ interface InputProps extends Omit<ComponentProps<'input'>, 'type'> {
   errorMessage?: string
   allowClear?: boolean
   classNames?: {
+    container?: string
     input?: string
     icon?: string
     errorMessage?: string
   }
+  onClear?: () => void
   colors?: {
     primary?: string
     error?: string
@@ -49,6 +51,7 @@ export function Input({
   className,
   classNames,
   ref,
+  onClear,
   ...props
 }: InputProps) {
   const [value, setValue] = useState(defaultValue || '')
@@ -58,14 +61,16 @@ export function Input({
   }
   const handleClear = () => {
     setValue('')
+    onClear?.()
   }
   const clearButtonVisible = value && !disabled && allowClear
 
   return (
     <Box
+      className={classNames?.container}
       display="inline-block"
       pos="relative"
-      selectors={{ '&,&>*': { boxSizing: 'border-box' } }}
+      selectors={{ '&, & *': { boxSizing: 'border-box' } }}
     >
       {icon && (
         <Center
@@ -102,7 +107,7 @@ export function Input({
           outline: 'none',
         }}
         _hover={{
-          border: '1px solid var(--primary, light-dark(red, blue))',
+          border: '1px solid var(--primary, light-dark(#674DC7, #8163E1))',
         }}
         _placeholder={{
           color: 'var(--inputPlaceholder, light-dark(#A9A8AB, #CBCBCB))',
@@ -141,7 +146,7 @@ export function Input({
         {...props}
       />
       {clearButtonVisible && <ClearButton onClick={handleClear} />}
-      {errorMessage && (
+      {error && errorMessage && (
         <Text
           aria-label="error-message"
           bottom="-8px"