+++ packages/components/src/index.ts
@@ -1,2 +1,18 @@
 export { Button } from './components/Button'
 export { Input } from './components/Input'
+export {
+  Select,
+  SelectContainer,
+  SelectDivider,
+  SelectOption,
+  SelectTrigger,
+  useSelect,
+} from './components/Select'
+export {
+  Stepper,
+  StepperContainer,
+  StepperDecreaseButton,
+  StepperIncreaseButton,
+  StepperInput,
+  useStepper,
+} from './components/Stepper'