+++ packages/components/CHANGELOG.md
@@ -1,5 +1,45 @@
 # @devup-ui/components
 
+## 0.1.8
+
+### Patch Changes
+
+- 3fd00d6: Fix SelectTrigger styleOrder
+
+## 0.1.7
+
+### Patch Changes
+
+- 10ed114: Fix exports. Fix Stepper.
+- fd19288: Fix exports
+
+## 0.1.6
+
+### Patch Changes
+
+- 0ef7821: Fix Input, Select, Stepper
+
+## 0.1.5
+
+### Patch Changes
+
+- fdaff93: Add Select component
+- 247f40a: Add Select component
+
+## 0.1.4
+
+### Patch Changes
+
+- 6948495: Add Stepper component
+- 6b51aa1: Fix Input test snapshot
+- 082e519: Fix Input
+
+## 0.1.3
+
+### Patch Changes
+
+- 9d7b342: Add onClear prop in input component
+
 ## 0.1.2
 
 ### Patch Changes