+++ packages/next-plugin/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @devup-ui/next-plugin
 
+## 1.0.29
+
+### Patch Changes
+
+- @devup-ui/webpack-plugin@1.0.28
+
 ## 1.0.28
 
 ### Patch Changes